halp: fix std::span narrowing in cpu_buffer_input::cast() on 32-bit
raw_buffer::byte_size is int64_t, so 'byte_size / sizeof(T)' stays 64-bit and
narrows to std::span's size_type in the braced initializer. On 64-bit size_t the
conversion is silent, but where size_t is 32-bit (wasm/emscripten) it fails to
compile under -Wc++11-narrowing. Cast the count to std::size_t explicitly.

+++ include/halp/buffer.hpp
@@ -65,13 +65,15 @@ struct cpu_buffer_input
   auto cast()
   {
     return std::span<T>{
-        reinterpret_cast<T*>(buffer.raw_data), buffer.byte_size / sizeof(T)};
+        reinterpret_cast<T*>(buffer.raw_data),
+        static_cast<std::size_t>(buffer.byte_size / sizeof(T))};
   }
   template <typename T>
   auto cast() const noexcept
   {
     return std::span<const T>{
-        reinterpret_cast<const T*>(buffer.raw_data), buffer.byte_size / sizeof(T)};
+        reinterpret_cast<const T*>(buffer.raw_data),
+        static_cast<std::size_t>(buffer.byte_size / sizeof(T))};
   }
 
   halp::raw_buffer buffer{};
